requirements.txt
# Core FastAPI dependencies
fastapi==0.111.*
uvicorn[standard]==0.29.*
pydantic==2.*
pydantic-settings==2.*

# NLP and ML libraries
spacy==3.7.*
sentence-transformers==2.*
torch>=2.0.0
transformers==4.*

# Text processing and utilities
nltk==3.*
textblob==0.17.*
langdetect==1.0.*
python-multipart==0.0.*

# Web scraping
requests==2.*
beautifulsoup4==4.*
selenium==4.*
aiohttp==3.*

# Database (optional)
sqlalchemy==2.*
alembic==1.*

# Caching and performance
redis==5.*
asyncio-throttle==1.*

# Development and testing
pytest==7.*
pytest-asyncio==0.23.*
black==23.*
flake8==6.*

# Logging and monitoring
structlog==23.*
python-json-logger==2.*

# Environment and configuration
python-dotenv==1.*


google-api-python-client
